requirements.txt
alembic==1.18.4
blinker==1.9.0
click==8.3.1
colorama==0.4.6
dnspython==2.8.0
email-validator==2.3.0
Flask==3.1.3
Flask-Login==0.6.3
Flask-Migrate==4.1.0
Flask-Principal==0.4.0
Flask-Security-Too==5.7.1
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
greenlet==3.3.2
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
libpass==1.9.3
Mako==1.3.10
MarkupSafe==3.0.3
mysql-connector-python==9.6.0
pymongo==4.16.0
PyMySQL==1.1.2
SQLAlchemy==2.0.48
tomli==2.4.1
typing_extensions==4.15.0
Werkzeug==3.1.6
WTForms==3.2.1
